Historical Clip Maps
These two maps are both clips of a study area. The clip map of California shows the location of hospitals and 3 specific historical hospitals in Orange County. The second clip map of Athens Georgia shows schools in the northern park of downtown with historical schools specified with a unique symbol.