Nighttime Emergency Response-Atlanta
This map shows the nighttime emergency response areas for hospitals within the city limit of Atlanta. In this map, methods like the creation of Thiessen Polygons was used to to find the closest hospital for any given location within the study area. The subjects of this map are displayed using dot density. Two additional charts are included to better understand the dat used to complete this map.